requirements.txt
pyodbc
xlsxwriter
typing
openpyxl 
cx_Oracle
python-dateutil
ConfigParser


